Fix test for real...
