+++ package.json
@@ -63,7 +63,9 @@
     "eslintConfig": {
         "extends": "react-app"
     },
-    "eslintIgnore": ["*.spec.ts"],
+    "eslintIgnore": [
+        "*.spec.ts"
+    ],
     "browserslist": [
         ">0.2%",
         "not dead",
